backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
statsmodels==0.14.6
python-dotenv==1.0.1
google-generativeai==0.7.2
groq==0.9.0
pydantic==2.7.1
python-multipart==0.0.9
